Updates to support Polar Stereographic CRS in Esri tools
This update alters the definitin of the polar stereographic coordinate reference system to be imported from WKT. This fixes an issue using derived outputs in Esri tools. For some reason, the Esri tools could not properly read the scale_factor projection parameter and thus the grids were not properly georeferenced. This addresses that shortcoming.