Added cleanup option to Build_Routing_Stack
Added the option to clean up temporary output files or not. It can be helpful for debugging issues to leave temporary output files in place, so now there is a global 'caleanUp' Boolean for selecting whether or not to remove these temporary files. Also an option was added so that the Examine_Outouts_of_GIS_Preprocessor.py script will not fail if a file does not have the `GeoTransform` variable attribute.