Update to solve zero_background channelgrid and stream order layers
There were some inconsistencies between the result of the channelgrid generation tools and the strahler stream order tools in Whitebox. Methods were added to specify zero_background conditions on the streams layer and set nodata values appropriately to screen out the background values. This was primarily a problem in reach-based routing configurations.