Adding NLINKS error checking and cosmetic changes
Added that ability to check a generated CHANNELGRID variable for channel connectivity errors, resolving the errors in the CHANNELGRID, FLOWDIRECTION, and STREAMORDER variables. This resolves an issue with WhiteBox tools giving results that cause WRF-hydro channel connectivity checks to fail. Also, print statements were evaluated to ensure consistent indentation in the output messaging of the tools.