Updating WhiteBox version for compatibility
Some issues with older WhiteBox Tools versions (2.2.0) seem to have caused incompatibility with the WRF-Hydro GIS Pre-processing tools due to the location where downloaded Whitebox binaries are extracted. The version 2.3.5 was tested to be compatible with the WRF-Hydro GIS Pre-processing tools, as it will extract binaries to the WBT folder.